following D&A's advice
the advice was:
The bear squash should have the beard's head lower bc atm the difference in nose height isn't significant enough. 
Add body follow through like Aaron blaise did in his walk cycle: the head hangs a bit as the body is going down, and the head stays down a bit as the body is pushing up.